add distributed metric
